feat(semantic): introduce intent definitions and resolution defaults
Made-with: Cursor
