+++ sharedPrompts/src/main/java/org/example/sharedprompts/domain/prompt/domain/semantic/IntentDefinitionEntriesProvider.java
@@ -1,0 +1,14 @@
+package org.example.sharedprompts.domain.prompt.domain.semantic;
+
+import java.util.List;
+
+/**
+ * Provides intent definition entries.
+ * <p>
+ * Data ownership lives in providers; {@link IntentDefinitionDataSource} only aggregates.
+ */
+interface IntentDefinitionEntriesProvider {
+
+    List<IntentDefinitionDataSource.IntentDefinitionEntry> entries();
+}
+
