+++ sharedPrompts/src/main/java/org/example/sharedprompts/domain/prompt/domain/semantic/IntentResolutionDefaultsEntriesProvider.java
@@ -1,0 +1,14 @@
+package org.example.sharedprompts.domain.prompt.domain.semantic;
+
+import java.util.List;
+
+/**
+ * Provides intent resolution-default entries (objective / output needs / response shape).
+ * <p>
+ * Data ownership lives in providers; {@link IntentDefinitionDataSource} only aggregates.
+ */
+interface IntentResolutionDefaultsEntriesProvider {
+
+    List<IntentDefinitionDataSource.IntentResolutionDefaultEntry> entries();
+}
+
